LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/selectors/user-invalid-expected.txt


Test form interactions (reset / submit):







PASS :user-invalid selector should respond to user action
PASS :user-invalid selector properly interacts with submit & reset buttons
PASS :user-error selector should not be supported
FAIL required-input: A required input or textarea should match :user-invalid if a user types into it and then clears it before blurring. assert_true: Element should match :user-invalid after typing text and deleting it. expected true got false
FAIL required-textarea: A required input or textarea should match :user-invalid if a user types into it and then clears it before blurring. assert_true: Element should match :user-invalid after typing text and deleting it. expected true got false
PASS A required checkbox should match :user-invalid if the user unchecks it and blurs.
FAIL A required date should match :user-invalid if the user unchecks it and blurs. assert_equals: Date input value should match the testdriver input. expected "1234-01-01" but got "1"
